move DAO to mt-data module
