+++ common/pom.xml
@@ -37,6 +37,10 @@
       <groupId>com.google.guava</groupId>
       <artifactId>guava</artifactId>
     </dependency>
+    <dependency>
+      <groupId>org.apache.commons</groupId>
+      <artifactId>commons-lang3</artifactId>
+    </dependency>
     <!-- enunciate -->
     <dependency>
       <groupId>com.webcohesion.enunciate</groupId>
