+++ mt-data/src/main/java/org/zanata/magpie/dao/AbstractDAO.java
