+++ mt-data/src/main/resources/META-INF/beans.xml
@@ -1,0 +1,6 @@
+<beans xmlns="http://java.sun.com/xml/ns/javaee"
+  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+  xmlns:weld="http://jboss.org/schema/weld/beans"
+  xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/beans_1_0.xsd">
+
+</beans>
