+++ server/src/main/java/org/zanata/magpie/api/service/impl/DocumentResourceImpl.java
@@ -26,7 +26,7 @@ import org.zanata.magpie.dao.LocaleDAO;
 import org.zanata.magpie.model.BackendID;
 import org.zanata.magpie.model.Document;
 import org.zanata.magpie.model.Locale;
-import org.zanata.magpie.service.DateRange;
+import org.zanata.magpie.dto.DateRange;
 import org.zanata.magpie.service.DocumentContentTranslatorService;
 import org.zanata.magpie.service.DocumentService;
 import org.zanata.magpie.util.UrlUtil;
